fix(review): extract shared franchise test helpers, guard requires_packs None, fix assertions
- Extract shared module-level constants and compute_food_truck_tool_surface
  from both test files into tests/franchise/_helpers.py to eliminate duplication
- Add `or []` guard to `sorted(recipe.requires_packs)` in franchise_runtime
  fixture to prevent TypeError when requires_packs is None
- Add `len(enable_calls) == 4` assertion before indexed access in
  test_food_truck_startup_enables_tags_in_order for cleaner failure output
- Rename test_food_truck_with_empty_requires_packs_sees_only_kitchen_core to
  test_food_truck_with_empty_requires_packs_sees_full_kitchen_fallback to
  accurately reflect the full-kitchen-fallback behavior (mcp.enable({"kitchen"}))
  triggered when AUTOSKILLIT_L2_TOOL_TAGS is empty
